Create Play Tic Tac Toe.py
The traditional two-player game Tic-Tac-Toe uses a nine-square grid. Each player takes turns marking their spot with an O or an X. The winner is the first to place three Os or Xs in a diagonal, horizontal, or vertical pattern. As each player tries to form a chain, they must also block their opponent.

Because it makes use of object-oriented programming, this is a unique and entertaining Python project for novices. 



This project will teach you one of the most important Python concepts, which is creating a new class named TicTacToe. The characteristics of the game will subsequently be represented using the class properties and methods.

Examine these techniques closely to see how we may package the different behaviours required to recreate this game in an organised manner using object-oriented programming.

Nested loops that check the grid's columns, rows, and diagonals for a winning state and the Python set data type, which is utilised to contain unique values, are two novel features of this beginner-friendly Python project idea. Now that you are more comfortable with it, this programme also chooses a random participant to begin the game using the Python random package.